% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docnload.R
\name{docnload}
\alias{docnload}
\title{Doc and Load}
\usage{
docnload()
}
\value{
Nothing
}
\description{
Convienience function to regularly run key dev functions
since buttons are no longer available in Postron.
}
