% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_comments.R
\name{get_column_comments}
\alias{get_column_comments}
\title{Get Column Comments}
\usage{
get_column_comments(db_connection)
}
\arguments{
\item{db_connection}{DBI::dbconnect object; A database connection.}
}
\value{
A data frame of column comments.
}
\description{
Get column comments.
}
