% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fk.R
\name{get_fk}
\alias{get_fk}
\title{Get Foreign Keys}
\usage{
get_fk(db_connection)
}
\arguments{
\item{db_connection}{DBI::dbconnect object; A database connection.}
}
\value{
A data frame.
}
\description{
Get foreign keys.
}
