% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pk.R
\name{get_pk}
\alias{get_pk}
\title{Get Primary Keys}
\usage{
get_pk(db_connection)
}
\arguments{
\item{db_connection}{DBI::dbconnect object; A database connection.}
}
\value{
A data frame list the primary key column for each table.
}
\description{
Get data base primary keys
}
