% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qri_connection.R
\name{get_qri_connection}
\alias{get_qri_connection}
\title{Get QRI Database Connection}
\usage{
get_qri_connection(set_key = TRUE)
}
\arguments{
\item{set_key}{boolean; Should the user be prompted to use the
keyring::key_set function to create the key_service on this
computer? If FALSE, uses the password store on the
computer.}
}
\value{
An OraConnection object whose class extends DBIConnection. This
object is used to execute SQL queries on the USACE QRI database.
}
\description{
Get a database connection to the USACE QRI schema. Only works
on the USACE network CorpsNet.
}
