% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_comments.R
\name{get_table_comments}
\alias{get_table_comments}
\title{Get Table Comments}
\usage{
get_table_comments(db_connection)
}
\arguments{
\item{db_connection}{DBI::dbconnect object; A database connection.}
}
\value{
A data frame of table comments.
}
\description{
Get table comments.
}
