% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_info.R
\name{get_table_info}
\alias{get_table_info}
\title{Get Table Info}
\usage{
get_table_info(db_connection, table_vector)
}
\arguments{
\item{db_connection}{DBI::dbconnect object; A database connection.}

\item{table_vector}{character; A character vector of table names as
returned by the get_tables function.}
}
\value{
A list of tables containing a data frame of columns names(for
each table).
}
\description{
Get the columns for each table.
}
