% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Get Tables}
\usage{
get_tables(db_connection, exclude_pattern = "(?!)")
}
\arguments{
\item{db_connection}{A DBI::dbconnect object; A database connection.}

\item{exclude_pattern}{character; A regex character string matching tables
names to exclude. The default includes all tables.}
}
\value{
A character vector of table names.
}
\description{
Retrieve a list of tables from the database connection.
}
