% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triples_df.R
\name{get_triples_df}
\alias{get_triples_df}
\title{Get Triples Data Frame}
\usage{
get_triples_df(graph)
}
\arguments{
\item{graph}{rdf object; An rdflib rdf object.}
}
\value{
A data.frame of the RDF Triplestore.
}
\description{
Convert a RDF Triplestore into a data.frame.
}
