% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_uri.R
\name{is_valid_uri}
\alias{is_valid_uri}
\title{Check for a Valid URI}
\usage{
is_valid_uri(uri_string)
}
\arguments{
\item{uri_string}{character; A URI string.}
}
\value{
boolean; TRUE if valid, FALSE if not valid
}
\description{
Check for a valid URI.
}
