% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fks_to_rdf.R
\name{map_fks_to_rdf}
\alias{map_fks_to_rdf}
\title{Map Foreign Keys to RDF}
\usage{
map_fks_to_rdf(graph, namespace, fk_df)
}
\arguments{
\item{graph}{rdf object; An rdflib rdf object.}

\item{namespace}{named character vector; A "namespace" or RDF
vocabulary is a collection of namespace
Internationalized Resource Identifiers (IRI)
used to construct URIs. Must be specified as a
named vector of the form:
namespace prefix = namespace IRI}

\item{fk_df}{data.frame; A df returned by the get_pk function.}
}
\value{
An rdflib::rdf graph object.
}
\description{
Map Foreign Keys to RDF.
}
