% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graph_from_ttl.R
\name{read_graph_from_ttl}
\alias{read_graph_from_ttl}
\title{Read Graph From Turtle}
\usage{
read_graph_from_ttl(ttl_path)
}
\arguments{
\item{ttl_path}{Path to the Turtle file}
}
\value{
An rdflib rdf graph
}
\description{
read_graph_from_ttl(): parse ttl -> rdflib graph
}
