% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri.R
\name{uri}
\alias{uri}
\title{Create Absolute Uniform Resource Identifier (URI)}
\usage{
uri(namespace, prefix, local)
}
\arguments{
\item{namespace}{named character vector; A "namespace" or RDF vocabulary
is a collection of namespace Internationalized
Resource Identifiers (IRI) used to construct URIs. Must be
specified as a named vector of the form:
namespace prefix = namespace IRI}

\item{prefix}{character; Nmaespace prefix}

\item{local}{character;}
}
\value{
a URI built from the
}
\description{
Constructs an absolute Uniform Resource Identifier (URI) from
the specified namespace.
}
\details{
See https://www.w3.org/TR/rdf11-concepts for definitions of these
terms.
}
