% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_table_to_rdf.R
\name{map_table_to_rdf}
\alias{map_table_to_rdf}
\title{Map Table to RDF}
\usage{
map_table_to_rdf(
  graph,
  namespace,
  table_name,
  cols_df,
  pk_df,
  tab_comments_df,
  col_comments_df
)
}
\arguments{
\item{graph}{rdf object; An rdflib rdf object.}

\item{namespace}{named character vector; A "namespace" or RDF
vocabulary is a collection of namespace
Internationalized Resource Identifiers (IRI)
used to construct URIs. Must be specified as a
named vector of the form:
namespace prefix = namespace IRI}

\item{table_name}{character; A named character vector of table names.}

\item{cols_df}{data.frame;}

\item{pk_df}{data.frame;}

\item{tab_comments_df}{data.frame;}

\item{col_comments_df}{data.frame;}
}
\value{
An rdf graph object.
}
\description{
Adds RDF triples to an existing RDF triplestore of table
definitions.
}
